requirements.txt
tensorflow
numpy
Pillow
requests
beautifulsoup4
python-chess
stockfish